Move constants.py to the right location
